rename some things and wire up UI, nothing tested yet
