+++ src/SharedUI/Pages/Sync.razor
@@ -15,9 +15,33 @@
 	</div>
 }
 
-<HxButton OnClick="SyncAsync" Color="ThemeColor.Primary" Enabled="@IsSyncButtonEnabled()">
-	Sync <span class="badge bg-secondary">@selectedItems.Count()</span>
-</HxButton>
+<HxCard>
+	<BodyTemplate>
+		<div class="row">
+			<div class="col-md-4">
+				<HxSwitch @bind-Value="forceStackWorkouts" Inline="true"> 
+					<TextTemplate>
+						Stack Workouts
+                        <HxPopover Trigger="PopoverTrigger.Hover|PopoverTrigger.Click|PopoverTrigger.Focus"
+                                   Title="<b>Stack Workouts</b>"
+                                   Content="@StackWorkoutsDocumentation"
+                                   Html="true">
+                            <HxIcon Icon="@BootstrapIcon.QuestionCircle" />
+                        </HxPopover>
+					</TextTemplate>
+				</HxSwitch>
+				
+			</div>
+			<div class="col-md-4">
+			</div>
+			<div class="col-md-4 text-end">
+				<HxButton OnClick="SyncAsync" Color="ThemeColor.Primary" Enabled="@IsSyncButtonEnabled()">
+					Sync <span class="badge bg-secondary">@selectedItems.Count()</span>
+				</HxButton>
+			</div>
+		</div>
+	</BodyTemplate>
+</HxCard>
 
 <HxGrid TItem="PelotonWorkout" MultiSelectionEnabled="true" Responsive="true" DataProvider="LoadDataAsync" @bind-SelectedDataItems="selectedItems" PageSize="PageSize" SelectionEnabled="false">
 	<Columns>
@@ -29,15 +53,12 @@
 	</Columns>
 </HxGrid>
 
-<HxButton OnClick="SyncAsync" Color="ThemeColor.Primary" Enabled="@IsSyncButtonEnabled()">
-	Sync <span class="badge bg-secondary">@selectedItems.Count()</span>
-</HxButton>
-
 @code {
 
 	private HashSet<PelotonWorkout> selectedItems = new();
 	private static int PageSize = 25;
 	private bool PelotonSettingNotConfiguredYet = false;
+	private bool forceStackWorkouts = false;
 
 	private GarminMfaModal? _garminMfaModal;
 
@@ -142,4 +163,6 @@
 	{
 		return selectedItems.Any();
 	}
+
+	private string StackWorkoutsDocumentation => $"Checking this will force the selected workouts to be combined into a single workout.  Selecting this will override anything you have configured in Settings.  It will stack the workouts regardless of how much time is between them so long as the workouts are of the same type..";
 }
