+++ src/SharedUI/Shared/AppSettingsForm.razor
@@ -13,7 +13,7 @@
 					Title="<b>Automatic Syncing</b>"
 					Content="@AutomaticSyncingDocumentation"
 					Html="true">
-						<HxBadge Type="BadgeType.RoundedPill" Color="ThemeColor.Info">?</HxBadge>
+						<HxIcon Icon="@BootstrapIcon.QuestionCircle" />
 					</HxPopover>
 				</HeaderTemplate>
 				<BodyTemplate>
