+++ src/SharedUI/Shared/GarminSettingsForm.razor
@@ -63,7 +63,7 @@
 									Title="<b>Upload Format</b>"
 									Content="@UploadFormatDocumentation"
 									Html="true">
-										<HxBadge Type="BadgeType.RoundedPill" Color="ThemeColor.Info">?</HxBadge>
+										<HxIcon Icon="@BootstrapIcon.QuestionCircle" />
 									</HxPopover>
 								</HeaderTemplate>
 								<BodyTemplate>
@@ -75,7 +75,7 @@
 								</BodyTemplate>
 							</HxCard>
 							<br />
-							<HxCard>
+							<HxCard CssClass="border-danger text-danger">
 								<HeaderTemplate>
 									Garmin Api Settings
 									<HxPopover Trigger="PopoverTrigger.Hover|PopoverTrigger.Click|PopoverTrigger.Focus"
